requirements.txt
discord.py==2.3.2
httpx==0.26.0
pytest==8.2.0
respx==0.21.1
